requirements.txt

fastapi
uvicorn
google-generativeai
PyMuPDF
requests
pydantic

